fix(macos): make h264/vp8 uint16 length fields explicit-cast
The compound (uint16 << 8) | uint16 expression promotes to int via
the usual arithmetic conversions, which Clang's
-Wimplicit-int-conversion flags as losing precision when assigned
back to uint16_t under -Werror.  Wrap the bit-OR with an explicit
static_cast<uint16_t> in h264.cpp (nalu_len) and vp8.cpp (pic16).

No behaviour change: the values still fit in 16 bits because each
operand was already masked to its byte-range (0x00..0xFF).

+++ src/modules/video_payload/src/h264.cpp
@@ -68,9 +68,9 @@ StapAResult parse_stap_a(core::ByteSpan payload) noexcept {
 
     std::size_t pos = 1;
     while (pos + 2 <= payload.size()) {
-        const std::uint16_t nalu_len =
+        const std::uint16_t nalu_len = static_cast<std::uint16_t>(
             (static_cast<std::uint16_t>(payload[pos])     << 8)
-          |  static_cast<std::uint16_t>(payload[pos + 1]);
+          |  static_cast<std::uint16_t>(payload[pos + 1]));
         pos += 2;
         if (nalu_len == 0 || pos + nalu_len > payload.size()) break;
